requirements.txt
fastapi[all]
openai>=1.55.0
pydantic-settings>=2.6.1
python-docx
lxml
python-dotenv
python-multipart
uvicorn
